refactor(npm): separate lockfile 3 detector
