+++ test/Microsoft.ComponentDetection.Detectors.Tests/NpmUtilitiesTests.cs
@@ -269,11 +269,8 @@ public class NpmUtilitiesTests
     {
         var testCases = new[]
         {
-            ("test", "test"),
-            ("@types/test", "@types/test"),
-            ("node_modules/test", "test"),
-            ("node_modules/@types/test", "@types/test"),
-            ("node_modules/root/node_modules/test", "test"),
+            ("test", "test"), ("@types/test", "@types/test"), ("node_modules/test", "test"),
+            ("node_modules/@types/test", "@types/test"), ("node_modules/root/node_modules/test", "test"),
             ("node_modules/root/node_modules/@types/test", "@types/test"),
             ("node_modules/rootA/node_modules/rootB/node_modules/test", "test"),
             ("node_modules/rootA/node_modules/rootB/node_modules/@types/test", "@types/test"),
@@ -282,32 +279,6 @@ public class NpmUtilitiesTests
         foreach (var (path, expectedModuleName) in testCases)
         {
             NpmComponentUtilities.GetModuleName(path).Should().Be(expectedModuleName);
-        }
-    }
-
-    [TestMethod]
-    public void TestNpmDetector_UpdateLockFileVersion()
-    {
-        var envMock = new Mock<IEnvironmentVariableService>();
-        var loggerMock = new Mock<ILogger>();
-
-        var testCases = new[]
-        {
-            new { version = 0, env = string.Empty, result = 0 }, new { version = 0, env = "true", result = 0 },
-            new { version = 1, env = string.Empty, result = 1 }, new { version = 1, env = "true", result = 1 },
-            new { version = 2, env = string.Empty, result = 2 }, new { version = 2, env = "true", result = 2 },
-
-            // 3 should be set as 2 if the env var is not set
-            new { version = 3, env = string.Empty, result = 2 }, new { version = 3, env = "true", result = 3 },
-        };
-
-        foreach (var testCase in testCases)
-        {
-            envMock
-                .Setup(x => x.GetEnvironmentVariable(NpmComponentUtilities.LockFile3EnvFlag))
-                .Returns(testCase.env);
-
-            NpmComponentUtilities.UpdateLockFileVersion(testCase.version, envMock.Object, loggerMock.Object).Should().Be(testCase.result);
         }
     }
 }
